bas-t2: IVariantMap be more strict. Convert datetimes.

+++ bas/layer-5/bas-repr/src/test/java/net/bodz/bas/repr/req/MutableRequestTest.java
@@ -15,7 +15,7 @@ public class MutableRequestTest
         assertEquals("mid", req.getAnchor());
         assertEquals(new PathEntries("def"), req.getPath());
         assertEquals("lenik", req.getString("name"));
-        assertEquals(13, req.getInt("age"));
+        assertEquals((Integer) 13, req.getInt("age"));
     }
 
 }
